server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.29.0",
    "fastapi>=0.109.0",
    "groq>=1.0.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.25",
    "uvicorn>=0.27.0",
]
